prov/shm: Separate SMR peers from SMR region
While SMR regions can map to endpoints, the data structure that holds
the list of SMR peers should be in the scope of domain. This would
allow nature mapping of the information to address vector and avoid
duplication for each endpoint.

A new object "SMR map" is introduced as the data base of all SMR peers.
This object is required as one of the parameters to create SMR regions.
Each SMR region refers to the SMR map to resolve the peers. The SMR map
is a local information and doesn't need to be part of the shared region.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>